+++ src/runtime/src/value.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub enum Value {
-    Number(f64),
+    Number(i128),
     String(String),
     Bool(bool),
     Null,
